courses_statistics/serializers.py
from rest_framework import serializers

from courses_statistics.models import UserStatistics
from usermanager.models import CustomUser


class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'


class UserStatisticsSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserStatistics
        fields = "__all__"
