requirements.txt
Django==4.2.6
django-cookie-consent==0.4.0
django-jazzmin==2.6.0
django-recaptcha==3.0.0
django-mathfilters==1.0.0
djangorestframework==3.14.0
Pillow==10.0.1
gunicorn==21.2.0